+++ CHANGES.md
@@ -1,3 +1,10 @@
+0.7.0
+-----
+* New timer subsystem! Bullet point by bullet point details below
+* Celluloid.sleep and Celluloid#sleep let an actor continue processing messages
+* Celluloid.receive and Celluloid#receive now accept an optional timeout
+* Celluloid::Mailbox#receive now accepts an optional timeout
+
 0.6.2
 -----
 * List all registered actors with Celluloid::Actor.registered